+++ superduperdb/cli/serve.py
@@ -4,8 +4,8 @@ import typing as t
 from superduperdb.datalayer.base.build import build_datalayer
 from superduperdb.datalayer.base.cdc import DatabaseWatcher
 from superduperdb.datalayer.mongodb.query import Collection
-from superduperdb.serve.server import serve as _serve
-from superduperdb.cluster.dask.dask_client import dask_client
+from superduperdb.cluster.server import serve as _serve
+from superduperdb.cluster.dask_client import dask_client
 from superduperdb import CFG
