+++ tests/integration/conftest.py
@@ -10,13 +10,13 @@ from pymongo import MongoClient
 from tenacity import RetryError, Retrying, stop_after_delay
 
 from superduperdb import CFG
-from superduperdb.cluster.dask.dask_client import dask_client
+from superduperdb.cluster.dask_client import dask_client
 from superduperdb.core.document import Document
 from superduperdb.datalayer.base.build import build_datalayer
 from superduperdb.encoders.torch.tensor import tensor
 from superduperdb.misc.config import DataLayer, DataLayers
 from superduperdb.models.torch.wrapper import TorchModel
-from superduperdb.serve.server import serve
+from superduperdb.cluster.server import serve
 
 '''
 All pytest fixtures with _package scope_ are defined in this module.
